Change balance and exchange rate to BigDecimal

+++ src/main/java/io/renatofreire/transaction_manager/dto/BankAccountDTO.java
@@ -1,9 +1,11 @@
 package io.renatofreire.transaction_manager.dto;
 
+import java.math.BigDecimal;
+
 public record BankAccountDTO(
         Long id,
         String currency,
-        Double balance,
+        BigDecimal balance,
         Long ownerId
 ) {
 }
